F13A-professionals-master/requirements.txt
flask
flask-cors
nltk
pillow
